Moved amqp into client package
