+++ client/amqp/amqp.go
