+++ client/amqp/amqp_test.go
