+++ client/amqp/option.go
