+++ client/amqp/option_test.go
