+++ examples/third/main.go
@@ -10,10 +10,10 @@ import (
 	"github.com/beatlabs/patron/async"
 	"github.com/beatlabs/patron/async/kafka"
 	"github.com/beatlabs/patron/async/kafka/group"
+	"github.com/beatlabs/patron/client/amqp"
 	"github.com/beatlabs/patron/encoding/json"
 	"github.com/beatlabs/patron/examples"
 	"github.com/beatlabs/patron/log"
-	"github.com/beatlabs/patron/trace/amqp"
 )
 
 const (
